fix: move edit video to video details
