+++ apps/videos-admin/src/app/[locale]/(dashboard)/videos/[videoId]/_VideoView/Metadata/StudyQuestions/StudyQuestions.tsx
