+++ apps/videos-admin/src/app/[locale]/(dashboard)/videos/[videoId]/_VideoView/Tabs/TabLabel/TabLabel.tsx
