+++ apps/videos-admin/src/app/[locale]/(dashboard)/videos/[videoId]/_VideoView/VideoView.tsx
@@ -2,28 +2,44 @@
 
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
+import Divider from '@mui/material/Divider'
 import Stack from '@mui/material/Stack'
+import Tab from '@mui/material/Tab'
+import Tabs from '@mui/material/Tabs'
 import Typography from '@mui/material/Typography'
-import Image from 'next/image'
-import { useParams, usePathname, useRouter } from 'next/navigation'
+import { useParams } from 'next/navigation'
 import { useTranslations } from 'next-intl'
-import { ReactElement } from 'react'
+import { ReactElement, SyntheticEvent, useState } from 'react'
 
 import { PublishedChip } from '../../../../../../components/PublishedChip'
 import { useAdminVideo } from '../../../../../../libs/useAdminVideo'
 
+import { Children } from './Children'
+import { Editions } from './Editions'
+import { Metadata } from './Metadata'
+import { Subtitles } from './Subtitles'
+import { TabContainer } from './Tabs/TabContainer'
+import { TabLabel } from './Tabs/TabLabel'
+import { Variants } from './Variants'
+
 export function VideoView(): ReactElement {
   const t = useTranslations()
   const params = useParams<{ videoId: string; locale: string }>()
-  const pathname = usePathname()
-  const router = useRouter()
+  const [tabValue, setTabValue] = useState(0)
+  const [isEdit, setIsEdit] = useState(false)
 
-  const { data } = useAdminVideo({
+  const { data, loading } = useAdminVideo({
     variables: { videoId: params?.videoId as string }
   })
 
+  const video = data?.adminVideo
+
   function handleEdit(): void {
-    router.push(`${pathname}/edit`)
+    setIsEdit(!isEdit)
+  }
+
+  function handleTabChange(e: SyntheticEvent, newValue: number): void {
+    setTabValue(newValue)
   }
 
   return (
@@ -38,6 +54,7 @@ export function VideoView(): ReactElement {
         flexWrap="wrap"
         sx={{ mb: 2, alignItems: 'center' }}
       >
+        {isEdit ? <Typography variant="h4">{t('Editing')} :</Typography> : null}
         <Typography variant="h4">{data?.adminVideo.title[0].value}</Typography>
         <PublishedChip published={data?.adminVideo.published ?? false} />
         <Button
@@ -46,36 +63,55 @@ export function VideoView(): ReactElement {
           size="small"
           sx={{ ml: 'auto', width: 'min-width' }}
         >
-          {t('Edit')}
+          {!isEdit ? t('Edit') : t('Cancel')}
         </Button>
       </Stack>
       <Stack gap={2} sx={{ flexDirection: { xs: 'column', sm: 'row' } }}>
-        <Box
-          sx={{
-            position: 'relative',
-            height: 225,
-            width: { xs: 'auto', sm: 400 },
-            borderRadius: 2,
-            overflow: 'hidden',
-            flexShrink: 0
-          }}
-        >
-          <Image
-            src={data?.adminVideo.images[0].mobileCinematicHigh as string}
-            alt={`${data?.adminVideo.imageAlt[0].value}`}
-            layout="fill"
-            objectFit="cover"
-            priority
-          />
-        </Box>
-        <Box sx={{ width: '100%' }}>
-          <Stack direction="column" sx={{ width: '100%' }}>
-            <Box sx={{ width: '100%' }}>
-              <Typography variant="body1">
-                {data?.adminVideo.description[0].value}
-              </Typography>
-            </Box>
-          </Stack>
+        <Box width="100%">
+          {loading ? (
+            <Typography>{t('Loading...')}</Typography>
+          ) : (
+            <>
+              <Tabs
+                value={tabValue}
+                onChange={handleTabChange}
+                aria-label="video-edit-tabs"
+              >
+                <Tab label={<TabLabel label="Metadata" />} />
+                <Tab
+                  label={
+                    <TabLabel
+                      label="Children"
+                      count={video?.children?.length}
+                    />
+                  }
+                />
+                <Tab
+                  label={
+                    <TabLabel
+                      label="Variants"
+                      count={video?.variants?.length}
+                    />
+                  }
+                />
+                <Tab label={<TabLabel label="Editions" />} />
+              </Tabs>
+              <Divider sx={{ mb: 4 }} />
+              <TabContainer value={tabValue} index={0}>
+                <Metadata video={video} loading={loading} />
+              </TabContainer>
+              <TabContainer value={tabValue} index={1}>
+                <Children childVideos={video?.children} />
+              </TabContainer>
+              <TabContainer value={tabValue} index={2}>
+                <Variants variants={video?.variants} />
+              </TabContainer>
+              <TabContainer value={tabValue} index={3}>
+                <Editions editions={[]} />
+                <Subtitles subtitles={video?.subtitles} videoId={video?.id} />
+              </TabContainer>
+            </>
+          )}
         </Box>
       </Stack>
     </Stack>
