+++ apps/videos-admin/src/app/[locale]/(dashboard)/videos/[videoId]/_VideoView/data.mock.ts
